requirements.txt
python-telegram-bot==21.5
anthropic>=0.40.0
requests==2.32.3
